feat(Guardian): APL update for Empowered Shapeshifting

+++ HeroRotation_Druid/Guardian.lua
@@ -289,16 +289,16 @@ local function Bear()
   if S.Mangle:IsCastable() and ((Player:BuffUp(S.Incarnation) and Enemies8yCount <= 4) or (Player:BuffUp(S.Incarnation) and S.SouloftheForest:IsAvailable() and Enemies8yCount <= 5) and ((Player:Rage() < 88) and Enemies8yCount < 11) or ((Player:Rage() < 83) and Enemies8yCount < 11 and S.SouloftheForest:IsAvailable())) then
     if Cast(S.Mangle, nil, nil, not IsInMeleeRange) then return "mangle bear 52"; end
   end
-  -- rip,if=buff.cat_form.up&(!dot.rip.ticking|refreshable)&combo_points>=3&active_enemies<3
-  if S.Rip:IsReady() and (Player:BuffUp(S.CatForm) and (Target:DebuffDown(S.RipDebuff) or Target:DebuffRefreshable(S.RipDebuff)) and Player:ComboPoints() >= 3 and Enemies8yCount < 3) then
+  -- rip,if=buff.cat_form.up&(!dot.rip.ticking|refreshable)&combo_points>=3&active_enemies<3&!talent.empowered_shapeshifting.enabled
+  if S.Rip:IsReady() and (Player:BuffUp(S.CatForm) and (Target:DebuffDown(S.RipDebuff) or Target:DebuffRefreshable(S.RipDebuff)) and Player:ComboPoints() >= 3 and Enemies8yCount < 3 and not S.EmpoweredShapeshifting:IsAvailable()) then
     if Cast(S.Rip, nil, nil, not IsInMeleeRange) then return "rip bear 54"; end
   end
-  -- ferocious_bite,if=buff.cat_form.up&dot.rip.ticking&combo_points>4&active_enemies<3
-  if S.FerociousBite:IsReady() and (Player:BuffUp(S.CatForm) and Target:DebuffUp(S.RipDebuff) and Player:ComboPoints() > 4 and Enemies8yCount < 3) then
+  -- ferocious_bite,if=buff.cat_form.up&dot.rip.ticking&combo_points>4&active_enemies<3&!talent.empowered_shapeshifting.enabled
+  if S.FerociousBite:IsReady() and (Player:BuffUp(S.CatForm) and Target:DebuffUp(S.RipDebuff) and Player:ComboPoints() > 4 and Enemies8yCount < 3 and not S.EmpoweredShapeshifting:IsAvailable()) then
     if Cast(S.FerociousBite, nil, nil, not IsInMeleeRange) then return "ferocious_bite bear 56"; end
   end
-  -- rake,if=talent.fluid_form.enabled&buff.bear_form.up&!buff.incarnation_guardian_of_ursoc.up&(refreshable|!dot.rake.ticking)&active_enemies<3&!talent.lunar_insight.enabled&energy>70|buff.cat_form.up&active_enemies<3&!talent.lunar_insight.enabled&talent.fluid_form.enabled&energy>70&(refreshable|!dot.rake.ticking)
-  if S.Rake:IsReady() and not IsTanking and (S.FluidForm:IsAvailable() and Player:BuffUp(S.BearForm) and Player:BuffDown(S.Incarnation) and (Target:DebuffRefreshable(S.RakeDebuff) or Target:DebuffDown(S.RakeDebuff)) and Enemies8yCount < 3 and not S.LunarInsight:IsAvailable() and Player:Energy() > 70 or Player:BuffUp(S.CatForm) and Enemies8yCount < 3 and not S.LunarInsight:IsAvailable() and S.FluidForm:IsAvailable() and Player:Energy() > 70 and (Target:DebuffRefreshable(S.RakeDebuff) or Target:DebuffDown(S.RakeDebuff))) then
+  -- rake,if=talent.fluid_form.enabled&buff.bear_form.up&!buff.incarnation_guardian_of_ursoc.up&(refreshable|!dot.rake.ticking)&active_enemies<3&!talent.lunar_insight.enabled&energy>70&!talent.empowered_shapeshifting.enabled|buff.cat_form.up&active_enemies<3&!talent.lunar_insight.enabled&talent.fluid_form.enabled&energy>70&(refreshable|!dot.rake.ticking)&!talent.empowered_shapeshifting.enabled
+  if S.Rake:IsReady() and not IsTanking and (S.FluidForm:IsAvailable() and Player:BuffUp(S.BearForm) and Player:BuffDown(S.Incarnation) and (Target:DebuffRefreshable(S.RakeDebuff) or Target:DebuffDown(S.RakeDebuff)) and Enemies8yCount < 3 and not S.LunarInsight:IsAvailable() and Player:Energy() > 70 and not S.EmpoweredShapeshifting:IsAvailable() or Player:BuffUp(S.CatForm) and Enemies8yCount < 3 and not S.LunarInsight:IsAvailable() and S.FluidForm:IsAvailable() and Player:Energy() > 70 and (Target:DebuffRefreshable(S.RakeDebuff) or Target:DebuffDown(S.RakeDebuff)) and not S.EmpoweredShapeshifting:IsAvailable()) then
     if Cast(S.Rake, nil, nil, not IsInMeleeRange) then return "rake bear 58"; end
   end
   -- thrash_bear,if=active_enemies>1
@@ -313,12 +313,12 @@ local function Bear()
   if S.ThrashBear:IsCastable() then
     if Cast(S.ThrashBear, nil, nil, not IsInAoERange) then return "thrash bear 64"; end
   end
-  -- rake,if=talent.fluid_form.enabled&buff.bear_form.up&!buff.incarnation_guardian_of_ursoc.up&(refreshable|!dot.rake.ticking)&active_enemies>3&!talent.lunar_insight.enabled&cooldown.mangle.remains<gcd
-  if S.Rake:IsReady() and not IsTanking and (S.FluidForm:IsAvailable() and Player:BuffUp(S.BearForm) and Player:BuffDown(S.Incarnation) and (Target:DebuffRefreshable(S.RakeDebuff) or Target:DebuffDown(S.RakeDebuff)) and Enemies8yCount > 3 and not S.LunarInsight:IsAvailable() and S.Mangle:CooldownRemains() < Player:GCD()) then
+  -- rake,if=talent.fluid_form.enabled&buff.bear_form.up&!buff.incarnation_guardian_of_ursoc.up&(refreshable|!dot.rake.ticking)&active_enemies>3&!talent.lunar_insight.enabled&cooldown.mangle.remains<gcd&!talent.empowered_shapeshifting.enabled
+  if S.Rake:IsReady() and not IsTanking and (S.FluidForm:IsAvailable() and Player:BuffUp(S.BearForm) and Player:BuffDown(S.Incarnation) and (Target:DebuffRefreshable(S.RakeDebuff) or Target:DebuffDown(S.RakeDebuff)) and Enemies8yCount > 3 and not S.LunarInsight:IsAvailable() and S.Mangle:CooldownRemains() < Player:GCD() and not S.EmpoweredShapeshifting:IsAvailable()) then
     if Cast(S.Rake, nil, nil, not IsInMeleeRange) then return "rake bear 66"; end
   end
-  -- shred,if=talent.fluid_form.enabled&buff.bear_form.up&!buff.incarnation_guardian_of_ursoc.up&dot.rake.ticking&active_enemies<3&!talent.lunar_insight.enabled&energy>70|buff.cat_form.up&active_enemies<3&!talent.lunar_insight.enabled&talent.fluid_form.enabled&energy>70
-  if S.Shred:IsReady() and not IsTanking and (S.FluidForm:IsAvailable() and Player:BuffUp(S.BearForm) and Player:BuffDown(S.Incarnation) and Target:DebuffUp(S.RakeDebuff) and Enemies8yCount < 3 and not S.LunarInsight:IsAvailable() and Player:Energy() > 70 or Player:BuffUp(S.CatForm) and Enemies8yCount < 3 and not S.LunarInsight:IsAvailable() and S.FluidForm:IsAvailable() and Player:Energy() > 70) then
+  -- shred,if=talent.fluid_form.enabled&buff.bear_form.up&!buff.incarnation_guardian_of_ursoc.up&dot.rake.ticking&active_enemies<3&!talent.lunar_insight.enabled&energy>70&!talent.empowered_shapeshifting.enabled|buff.cat_form.up&active_enemies<3&!talent.lunar_insight.enabled&talent.fluid_form.enabled&energy>70&!talent.empowered_shapeshifting.enabled
+  if S.Shred:IsReady() and not IsTanking and (S.FluidForm:IsAvailable() and Player:BuffUp(S.BearForm) and Player:BuffDown(S.Incarnation) and Target:DebuffUp(S.RakeDebuff) and Enemies8yCount < 3 and not S.LunarInsight:IsAvailable() and Player:Energy() > 70 and not S.EmpoweredShapeshifting:IsAvailable() or Player:BuffUp(S.CatForm) and Enemies8yCount < 3 and not S.LunarInsight:IsAvailable() and S.FluidForm:IsAvailable() and Player:Energy() > 70 and not S.EmpoweredShapeshifting:IsAvailable()) then
     if Cast(S.Shred, nil, nil, not IsInMeleeRange) then return "shred bear 68"; end
   end
   -- moonfire,if=buff.galactic_guardian.up&buff.bear_form.up&talent.boundless_moonlight.enabled
